Split action servers and clients into separate modules
