+++ rclrs/src/action/server.rs
@@ -1,0 +1,130 @@
+use crate::{error::ToResult, rcl_bindings::*, Clock, Node, RclrsError, ENTITY_LIFECYCLE_MUTEX};
+use std::{
+    ffi::CString,
+    marker::PhantomData,
+    sync::{atomic::AtomicBool, Arc, Mutex, MutexGuard},
+};
+
+// SAFETY: The functions accessing this type, including drop(), shouldn't care about the thread
+// they are running in. Therefore, this type can be safely sent to another thread.
+unsafe impl Send for rcl_action_server_t {}
+
+/// Manage the lifecycle of an `rcl_action_server_t`, including managing its dependencies
+/// on `rcl_node_t` and `rcl_context_t` by ensuring that these dependencies are
+/// [dropped after][1] the `rcl_action_server_t`.
+///
+/// [1]: <https://doc.rust-lang.org/reference/destructors.html>
+pub struct ActionServerHandle {
+    rcl_action_server: Mutex<rcl_action_server_t>,
+    node: Node,
+    pub(crate) in_use_by_wait_set: Arc<AtomicBool>,
+}
+
+impl ActionServerHandle {
+    pub(crate) fn lock(&self) -> MutexGuard<rcl_action_server_t> {
+        self.rcl_action_server.lock().unwrap()
+    }
+}
+
+impl Drop for ActionServerHandle {
+    fn drop(&mut self) {
+        let rcl_action_server = self.rcl_action_server.get_mut().unwrap();
+        let mut rcl_node = self.node.handle.rcl_node.lock().unwrap();
+        let _lifecycle_lock = ENTITY_LIFECYCLE_MUTEX.lock().unwrap();
+        // SAFETY: The entity lifecycle mutex is locked to protect against the risk of
+        // global variables in the rmw implementation being unsafely modified during cleanup.
+        unsafe {
+            rcl_action_server_fini(rcl_action_server, &mut *rcl_node);
+        }
+    }
+}
+
+/// Trait to be implemented by concrete ActionServer structs.
+///
+/// See [`ActionServer<T>`] for an example
+pub trait ActionServerBase: Send + Sync {
+    /// Internal function to get a reference to the `rcl` handle.
+    fn handle(&self) -> &ActionServerHandle;
+    // /// Tries to take a new request and run the callback with it.
+    // fn execute(&self) -> Result<(), RclrsError>;
+}
+
+pub struct ActionServer<T>
+where
+    T: rosidl_runtime_rs::Action,
+{
+    _marker: PhantomData<fn() -> T>,
+    pub(crate) handle: Arc<ActionServerHandle>,
+    // goal_callback: (),
+    // cancel_callback: (),
+    // accepted_callback: (),
+}
+
+impl<T> ActionServer<T>
+where
+    T: rosidl_runtime_rs::Action,
+{
+    /// Creates a new action server.
+    pub(crate) fn new(node: &Node, clock: Clock, topic: &str) -> Result<Self, RclrsError>
+    where
+        T: rosidl_runtime_rs::Action,
+    {
+        // SAFETY: Getting a zero-initialized value is always safe.
+        let mut rcl_action_server = unsafe { rcl_action_get_zero_initialized_server() };
+        let type_support = <T as rosidl_runtime_rs::Action>::get_type_support()
+            as *const rosidl_action_type_support_t;
+        let topic_c_string = CString::new(topic).map_err(|err| RclrsError::StringContainsNul {
+            err,
+            s: topic.into(),
+        })?;
+
+        // SAFETY: No preconditions for this function.
+        let action_server_options = unsafe { rcl_action_server_get_default_options() };
+
+        {
+            let mut rcl_node = node.handle.rcl_node.lock().unwrap();
+            let rcl_clock = clock.rcl_clock();
+            let mut rcl_clock = rcl_clock.lock().unwrap();
+            let _lifecycle_lock = ENTITY_LIFECYCLE_MUTEX.lock().unwrap();
+
+            // SAFETY:
+            // * The rcl_action_server is zero-initialized as mandated by this function.
+            // * The rcl_node is kept alive by the Node because it is a dependency of the action server.
+            // * The topic name and the options are copied by this function, so they can be dropped
+            //   afterwards.
+            // * The entity lifecycle mutex is locked to protect against the risk of global
+            //   variables in the rmw implementation being unsafely modified during initialization.
+            unsafe {
+                rcl_action_server_init(
+                    &mut rcl_action_server,
+                    &mut *rcl_node,
+                    &mut *rcl_clock,
+                    type_support,
+                    topic_c_string.as_ptr(),
+                    &action_server_options,
+                )
+                .ok()?;
+            }
+        }
+
+        let handle = Arc::new(ActionServerHandle {
+            rcl_action_server: Mutex::new(rcl_action_server),
+            node: node.clone(),
+            in_use_by_wait_set: Arc::new(AtomicBool::new(false)),
+        });
+
+        Ok(Self {
+            _marker: Default::default(),
+            handle,
+        })
+    }
+}
+
+impl<T> ActionServerBase for ActionServer<T>
+where
+    T: rosidl_runtime_rs::Action,
+{
+    fn handle(&self) -> &ActionServerHandle {
+        &self.handle
+    }
+}
